cli: consolidate Cloud capabilities, servers, and sessions
