+++ .changeset/cli-cloud-consolidate.md
@@ -1,0 +1,5 @@
+---
+"@mcpjam/cli": major
+---
+
+Consolidate Cloud commands: `mcpjam cloud capabilities` is now `mcpjam cloud projects capabilities`; `projects server` is no longer an alias of `projects servers`; `chat-sessions` is removed in favor of `mcpjam cloud sessions list`.
